% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{biodbNcbi-package}
\alias{biodbNcbi}
\alias{biodbNcbi-package}
\title{biodbNcbi: biodbNcbi, a library for connecting to NCBI Databases.}
\description{
The biodbNcbi library is an extension of the biodb framework package, that provides access to NCBI databases CCDS, Gene, Pubchem Comp and Pubchem Subst. It allows to retrieve entries by their accession number, and run specific web services.
}
\details{
See vignette biodbNcbi:\preformatted{vignette('biodbNcbi', package='biodbNcbi')
}
}
\seealso{
\link{NcbiGeneConn}, \link{NcbiCcdsConn},
\link{NcbiPubchemCompConn} and \link{NcbiPubchemSubstConn}.
}
\author{
\strong{Maintainer}: Pierrick Roger \email{pierrick.roger@cea.fr}

}
