% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{biodbNcbi-package}
\alias{biodbNcbi}
\alias{biodbNcbi-package}
\title{biodbNcbi: biodbNcbi, a library for connecting to NCBI Databases.}
\description{
The biodbNcbi library provides access to the NCBI databases CCDS, Gene, Pubchem Comp and Pubchem Subst, using biodb package framework. It allows to retrieve entries by their accession number. Web services can be accessed for searching the database by name or mass.
}
\details{
See vignette biodbNcbi:\preformatted{vignette('biodbNcbi', package='biodbNcbi')
}
}
\seealso{
\link{NcbiGeneConn}, \link{NcbiCcdsConn},
\link{NcbiPubchemCompConn} and \link{NcbiPubchemSubstConn}.
}
\author{
\strong{Maintainer}: Pierrick Roger \email{pierrick.roger@cea.fr}

}
